fix mobile catalog closing after click
