+++ src/components/CatalogComponents/CatalogMenu/style.module.scss
