+++ src/components/MobileMenu/MobileSubCatalog/style.module.scss
@@ -38,3 +38,9 @@
 .menu_wrap span {
   width: 100%;
 }
+
+.child2_wrap {
+  position: absolute;
+  top: 0;
+  left: 0;
+}
