+++ src/components/MobileMenu/SubCatChildren/SubCatChildrenSecond.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+import styles from "./style.module.scss";
+
+export default function SubCatChildrenSecond() {
+  return <div className={styles.child2_wrap}>SubCatChildrenSecond</div>;
+}
